requirements.txt
langchain
langchain-core
langchain-openai
langchain-community

langserve
fastapi
uvicorn

pydantic
python-dotenv
sse_starlette

faiss-cpu
pypdf
pypdf2
pdfplumber

azure-storage-blob

psycopg_pool
psycopg

python-multipart